+++ biofefi/services/feature_importance/global_methods.py
@@ -1,7 +1,7 @@
 from typing import Any
-
 import pandas as pd
 import shap
+from sklearn.inspection import permutation_importance
 
 from biofefi.utils.logging_utils import Logger
 
@@ -44,37 +44,49 @@ def calculate_global_shap_values(
     return shap_df, shap_values
 
 
-def calculate_local_shap_values(
+def calculate_permutation_importance(
     model,
     X: pd.DataFrame,
-    shap_reduce_data: int,
+    y: pd.Series,
+    permutation_importance_scoring: str,
+    permutation_importance_repeat: int,
+    random_state: int,
     logger: Logger,
-) -> tuple[pd.DataFrame, Any]:
-    """Calculate local SHAP values for a given model and dataset.
+):
+    """Calculate permutation importance for a given model and dataset.
 
     Args:
         model: Model object.
-        X (pd.DataFrame): The dataset.
-        shap_reduce_data (int): The percentage of data to use for SHAP calculation.
+        X (pd.DataFrame): Input features.
+        y (pd.Series): Target variable.
+        permutation_importance_scoring (str): Permutation importance scoring method.
+        permutation_importance_repeat (int): Number of repeats for importance scoring.
+        random_state (int): Seed for the random state.
         logger (Logger): The logger.
 
     Returns:
-        tuple[pd.DataFrame, Any]: SHAP dataframe and SHAP values.
+        permutation_importance: Permutation importance values
     """
-    logger.info(f"Calculating SHAP Importance for {model.__class__.__name__} model..")
 
-    if shap_reduce_data == 100:
-        explainer = shap.Explainer(model.predict, X)
-    else:
-        X_reduced = shap.utils.sample(X, int(X.shape[0] * shap_reduce_data / 100))
-        explainer = shap.Explainer(model.predict, X_reduced)
+    logger.info(
+        f"Calculating Permutation Importance for {model.__class__.__name__} model.."
+    )
 
-    shap_values = explainer(X)
+    # Use permutation importance in sklearn.inspection to calculate feature importance
+    permutation_importance_results = permutation_importance(
+        model,
+        X=X,
+        y=y,
+        scoring=permutation_importance_scoring,
+        n_repeats=permutation_importance_repeat,
+        random_state=random_state,
+    )
+    # Create a DataFrame with the results
+    permutation_importance_df = pd.DataFrame(
+        permutation_importance_results.importances_mean, index=X.columns
+    )
 
-    shap_df = pd.DataFrame(shap_values.values, columns=X.columns, index=X.index)
-    # TODO: scale coefficients between 0 and +1 (low to high impact)
-
-    logger.info("SHAP Importance Analysis Completed..")
+    logger.info("Permutation Importance Analysis Completed..")
 
     # Return the DataFrame
-    return shap_df, shap_values
+    return permutation_importance_df
